+++ assets/css/index.ts
@@ -1,5 +1,5 @@
 import "./primercss.scss";
 import "./actionlist/index.scss";
-import "dialogic-js/dialogic-js.css";
+import "./prompt.css";
 import "./primercss-overrides.css";
 import "./toggle-switch.css";