+++ assets/css/prompt.ts
@@ -1,1 +1,0 @@
-import "dialogic-js/dialogic-js.css";