+++ assets/index.ts
@@ -1,2 +1,1 @@
-export * from './js';
-export * from './css';
+export * from "./css";